Slowly moving to async first, small bug fixes for folder path
